feat(examples): add `r/morgan/{home,guestbook}` (#2345)
I wanted a simple realm that allows any user to interact with it and see
something happen. This is a guestbook realm; it allows anyone to leave a
small message, which will be associated with their address.

It also is a nice showcase of how to effectively do pagination with AVL
trees. I don't think we have many such examples so far.