feat(stdlibs): add strings.Replacer (#2816)
<!-- please provide a detailed description of the changes made in this
pull request. -->
prerequisite of #2802 

This pull request ports the files:
- replace.go
- replace_test.go

from the Golang standard library.

I added some tags on the code with the hope it will help to review the
code and to launch discussion if neccessary.
I could after remove these changes
```go
// Custom code: XXX_Some_Explanation
( code not present on the original go file)
.
.
.
// End of custom code
```

 

<details><summary>Contributors' checklist...</summary>

- [ ] Added new tests, or not needed, or not feasible
- [ ] Provided an example (e.g. screenshot) to aid review or the PR is
self-explanatory
- [ ] Updated the official documentation or not needed
- [ ] No breaking changes were made, or a `BREAKING CHANGE: xxx` message
was included in the description
- [ ] Added references to related issues and PRs
- [ ] Provided any useful hints for running manual tests
- [ ] Added new benchmarks to [generated
graphs](https://gnoland.github.io/benchmarks), if any. More info
[here](https://github.com/gnolang/gno/blob/master/.benchmarks/README.md).
</details>