wip: drafting possible patterns and implementation
This changes are just one idea in case we want to consider using a
single type, like Post or Node for example, for all board types.

The proposed ideas here are based on points discussed by the team up to
this point of the definition.

The idea is to see if we can define a good pattern to implement the
boards (boards, posts, comments, ...) using a single type as mentioned
by Jae at some point during early discussions.

These changes are NOT finished, they are a "showcase" for further
exploration and discussion at this point.