chore: fix issues reported by `go vet` (#2952)
<!-- please provide a detailed description of the changes made in this
pull request. -->
Passing `go vet` is the minimum level of code quality for a go project.
This is addressed here for the full mono-repo. Mainly trivial changes,
except for a few `copy lock` issues which could be more meaningful (kept
for a separate pull request).

Addresses #2954 (but not sufficient to close it yet).

Hint for reviewers: it's easier to review each commit individually.
 
<details><summary>Contributors' checklist...</summary>

- [x] Added new tests, or not needed, or not feasible
- [x] Provided an example (e.g. screenshot) to aid review or the PR is
self-explanatory
- [x] Updated the official documentation or not needed
- [x] No breaking changes were made, or a `BREAKING CHANGE: xxx` message
was included in the description
- [ ] Added references to related issues and PRs
- [x] Provided any useful hints for running manual tests
- [ ] Added new benchmarks to [generated
graphs](https://gnoland.github.io/benchmarks), if any. More info
[here](https://github.com/gnolang/gno/blob/master/.benchmarks/README.md).
</details>