ci: remove unused goreleaser files, ignore chain/ tags (#3004)
I don't know how to test this, but I think this may fix it. @ajnavarro
do you have a separate repo / access to the pro token and can see if
this configuration works?

I removed the other config files as they weren't used, and the
`nightly:` section as I don't think it's parsed, anyway: it doesn't
exist in the goreleaser schema.