feat: add `p/moul/txlink` + `p/moul/helplink` (#2887)
This PR aimed to promote the use of a `p/` library for managing special
help links from contracts.

It also provided an opportunity for me to realize that our discussion
about changing the `$` symbol would require some parsing and detection
from the `gnoweb` perspective. If we want a simple library like this
one, the goal should be to ideally craft a link to the current package
without specifying the realm path. Relative URLs worked well with `?`,
but they won't function with `$`.

As an alternative, we can have this package look for
`std.PrevRealm().PkgAddr` if it is not specified.

cc @jeronimoalbi @thehowl @leohhhn 

Related with #2602
Related with #2876

---------

Signed-off-by: moul <94029+moul@users.noreply.github.com>
Co-authored-by: Leon Hudak <33522493+leohhhn@users.noreply.github.com>