ci: benchmark only BenchmarkBenchdata (#3007)
A bit radical, but I'm open to other benchmarks we should include.

Essentially, in an effort to have a small amount of meaningful
benchmarks, I'd like for these to only be those in BenchmarkBenchdata.
Yes, I recognize this is tooting my own horn, but I think they are good
benchmarks that tell us, overall, if the GnoVM on a few reference
programs got slower or faster, and I found them useful in the past while
doing manual execution.

Most other benchmarks are micro-benchmarks, which aren't likely to
change often or to give us useful insight. I'm open to suggestions for
others that make sense to be tracked, but I think it's for the better if
we keep the number low so the CI for benchmarks can run in just a few
minutes.