chore: remove boardsv2 experiment drafts
We won't continue in that direction right now.
The current objective is to finish new features for `boards` which will
become `boards2`.

Previous experiments were aiming to have a "plugin" based approach
focused on flexibility but at this point we want to focus on a simpler
and straight forward approach based on extending the current `boards`
implementation.