feat: add `assertIsBoardName()` function
Also changed the expression to validate board names to check that the
name starts with three letters and ends with three numbers.

This has been defined this way to match `users` realm implementation.
See: https://github.com/gnolang/gno/issues/2827

In a next iteration we can support vanity names for boards, once vanity
names are supported for the `users` realm.