feat(examples): add haystack package and realm (#3082)
Haystack is a permissionless , immutable, content-addressed,
append-only, fixed-length key-value store for small payloads.

This is an experiment to port over a storage and validation
implementation of a tool I'd written several years ago called
[haystack](https://github.com/nomasters/haystack). My goal in porting
this over to gno is to provide a simple, correct, and test covered
implementation that is composable with other tools I plan to port over
as well.

## Overview

You store a needle in the haystack.

A Needle is 192 bytes. It is composed of 32 bytes for a sha256 hash, and
the 160 byte fixed-length payload.

```
hash     | payload
---------|----------
32 bytes | 160 bytes
```

The Haystack storage server supports two calls, other than Render, you
may "Add" a needle, by its hex-encoded string, or you may "Get" a needle
by its hex encoded hash. The add operation ensures that the needle is
valid and that it has not been added to the storage before. The Get
operation will return the full hex encoded needle if the hash exists in
the database, otherwise it will panic.

The structure is broken down into 2 packages and 1 very simple realm

### packages
- https://gno.land/p/demo/haystack/needle/
- https://gno.land/p/demo/haystack/

### realm
- https://gno.land/r/demo/haystack

## How to Try it out?

You can generate your own synthetic needle from the CLI by using this
magical one-liner.

```shell
➜  ~ (dd if=/dev/urandom bs=160 count=1 2>/dev/null | tee >(sha256sum |  cut -d' ' -f1) | od -An -t x1) | tr -d ' \n'
5d82091003a6749b46a96c38b2597ca96e9b0b272594249099dcf2ade188346679ca753999661820dad7beb351559c89a275ed4935a82245fda290906670ec7535b0b856dfccadd62e5f5399892455d2b524724ffdef8e58be03e9da4762c6ab582ce91c29a9e26ea9cc38b66953fdc425ad37baeb12c712e049ae6d456e682b6b63eea74ebf7a9d506ba486d08c9c54c5161d38a7fbc5fcbb1cdac370682ad6a59579167fd1aa1cd1fc109660a7eba36775d6b06058d72aa57debe63d0144b8
```
This leverages `dd`, `/dev/urandom`, `tee`, `cut`, `od`, and `tr` to
generate a properly formatted hex-encoded needle.

### Getting a needle

I've already stored the above needle in Haystack, so you can read it by
running this command from the CLI

```shell
gnokey maketx call -pkgpath "gno.land/r/demo/haystack" \
    -func "Get" \
    -gas-fee 1000000ugnot \
    -gas-wanted 2000000 \
    -send "" \
    -broadcast \
    -chainid "portal-loop" \
    -args "5d82091003a6749b46a96c38b2597ca96e9b0b272594249099dcf2ade1883466" \
    -remote "https://rpc.gno.land:443" \
    $YOUR_WALLET_ADDRESS
```



<details><summary>Contributors' checklist...</summary>

- [X] Added new tests, or not needed, or not feasible
- [X] Provided an example (e.g. screenshot) to aid review or the PR is
self-explanatory
- [X] Updated the official documentation or not needed
- [X] No breaking changes were made, or a `BREAKING CHANGE: xxx` message
was included in the description
- [X] Added references to related issues and PRs
- [X] Provided any useful hints for running manual tests
</details>