feat: add premission support to other "sync" public realm functions
There are function that require checks that should be done by an
implementer, like a DAO. Initially the implementer might choose to use
proposals under the hood making it impossible to return the ID of a
newly created entity syncronically. This change just adds permission
support but the pattern must be improved to handle these cases.